move notification settings into its own component
