+++ src/models/task.ts
@@ -3,6 +3,7 @@ import { Label } from './label'
 import { ColorPaletteProp } from '@mui/joy'
 import { dayOfMonthSuffix } from '../utils/date'
 import { IntervalUnit } from '@/utils/recurrance'
+import { Notification } from '@/models/notifications'
 
 export type RepeatOnce = {
   type: 'once'
@@ -50,16 +51,6 @@ export type RepeatDayOfTheMonths = {
 
 export type RepeatCustom = RepeatInterval | RepeatDaysOfTheWeek | RepeatDayOfTheMonths
 export type Frequency = RepeatOnce | RepeatDaily | RepeatWeekly | RepeatMonthly | RepeatYearly | RepeatCustom
-
-export type NotificationDisabled = {
-  enabled: false
-}
-export type NotificationMqtt = object
-export type NotificationTrigger = 'due_date' | 'pre_due' | 'overdue' | 'nag'
-export type NotificationEnabled = {
-  enabled: true
-} & Record<NotificationTrigger, boolean> & (NotificationMqtt)
-export type Notification = NotificationDisabled | NotificationEnabled
 
 export interface Task {
   id: string
